Dealing with numbers 20-29 in number_words
simple cases are up to 1-19, after that we can combine cases. So if
number is 20 or more we check to see if we need to combine the 2 words
together (20 is just 'twenty' with no trailing part). Using integer
division of 2y / 10 to give 2.

The remainder of 2y % 10 to calculate the trailing part.

Refactor: building up strings using concatination to mean only 1 return
location