number_words up to 999
Code was nearly working for all numbers up to 999 however it only worked
out 'hundred' string for 100-199 due to the calculation for this. A
simple solution is just to set this block to always return the 'hundred'
string by hard coding this. We are in the >= 100 so its okay.

Refactor: conditionals to be explitive on the boundaries.

Reactor: moved all strings (apart from white space) into words, this
includes 'and'.