Dealing with the variation of thousands in number_words
Deaing with the variation of 'thousand and X' vs 'thousand, X'. A sub
clause added to the non round thousand logic to deal with if its
"X, Y", when the number has a non zero digit in the hundreds column.
or if it has a zero digit in the hundreds column use "and".